refactor(store): allow adding padding to store
This allows aligning components across stores, to be able to find all
components belonging to a single entity.

All stores together form a grid of entity/component relations:

              Entity  1   2   3   4   5

    Position store:  10   5   x   0  10
    Velocity store:   x   1   x   x  10
    Mass store:       3   3   3   x  10

This shows how the element positions in the stores correspond to a
single entity. The first element in all stores contains the components
belonging to entity 1. Since entity 1 has no Velocity component, padding
is added to make this clear to the ECS system.

If a `System` wants to iterate over all entities containing `Position`
and `Velocity` components, the library will iterate over those two
stores, and filter out any position where either one of the two values
is `None`.

I'm convinced this is a very inefficient way to handle storing/
retrieving of components, but it's a good starting point on top of which
all other features can be built. Once the API stabilises, focus will
shift to improving the store implementation, and abstracting it into a
trait, to allow different store implementations with their own trade-
offs in terms of read/write performance and cache efficiency.

NOTE: the current implementation requires nightly, due to the use of
      `resize_default`. This will be changed in the future to support
      Rust stable.