feat(entity): support multiple components of same type
By not only keeping track of the _position_ in the store, but also the
maximum _length_ of the stored components for a single entity, we can
support storing multiple components of the same type.

For example:

              Entity  1   x   2   3

    Position store:  10   5   5   x
    Velocity store:   x   x   3   2
    Mass store:       3   x   x   8

This shows how Entity 1 has two `Position` components, with values 10
and 5, and one `Mass` component with value 3. The next entity (2) has
one `Position` component and one `Velocity` component.

In the future, we'll have to do some benchmarking to see what the trade-
offs are in terms of performance. Once the ECS is used in Rusty Rockets,
it might turn out that multiple components of the same type aren't a
useful feature, and we rather have some extra performance.

Or, perhaps this can be abstracted away so that the implementation of
the store determines if multiple components of the same type per entity
are allowed.

Either way, to be continued...