feat(system): add initial System implementation
This is a first working version of the `System` implementation that
allows querying exactly _ two read-only_ components.

A trivial example:

    use things::{Join, Query, Read, System, Things};

    struct MyTestSystem;
    impl<'a> System<'a> for MyTestSystem {
        type Query = Join<(Read<usize>, Read<String>)>;

        fn update(components: <Self::Query as Query<'a>>::Iter) {
            for (int, string) in components {
                assert_eq!(int, &10);
                assert_eq!(string, &"hello");
            }
        }
    }

    #[test]
    fn test_system() {
        let mut ecs = Things::new();
        ecs.create_entity((10, "hello"));
        ecs.execute_system::<MyTestSystem>();
    }

There is still plenty to improve:

* Improve type ergonomics such as having to specify the lifetime
  parameter `'a`.

* Having to call the `update` method with the
  `<Self::Query as Query<'a>>::Iter` type signature.

* Having to wrap the query in the `Join` struct.

And there's plenty of missing features that still need to be added:

* Access mutable components using `Write`.

* Execute ad-hoc system logic without having to wrap it in a struct.

* Have Things provide a scheduler so you don't have to manually trigger
  every System using `execute_system`.

* Being able to store arbitrary (mutable) data on systems to use in the
  `update` call.

* Extend system support to anything other than just an entity of two
  components.

And then there's also still plenty to optimize on the back end of the
implementation, such as performance improvements, better code
readability and a more SOLID approach to some of the current design
decisions. And most important of all (and more doable now that there's a
more solid foundation); adding unit tests and restructure code that
turns out to be hard to validate in a test.

But, this is another step in the right direction, and exciting to see
working.