feat(store): use internal mutability of components
The store now wraps the `Vec` of `Component`s in an `UnsafeCell`, to
allow accessing components mutably, without requiring the store itself
to be accessed mutably.

To guarantee correct access patterns, similar to Rust's compile-time
borrow-checker, a new "runtime borrow-checker" will have to be
implemented. This is NOT implemented yet in this commit, and thus the
implementation is unsound.

Before the 1.0 release, the runtime borrow-checker will be added.